+++ bankapplication/src/main/java/com/cognizant/bankapplication/repository/AccountRepository.java
@@ -1,10 +1,12 @@
 package com.cognizant.bankapplication.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cognizant.bankapplication.model.Account;
 
-public interface AccountRepository extends JpaRepository<Account, Long>{
+public interface AccountRepository extends JpaRepository<Account, BigInteger>{
 	
 	public Account findByAccountId(Long accountId); 
 }
