+++ bankapplication/src/main/java/com/cognizant/bankapplication/repository/TransactionRepository.java
@@ -1,11 +1,13 @@
 package com.cognizant.bankapplication.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cognizant.bankapplication.model.Transaction;
 
 public interface TransactionRepository extends JpaRepository<Transaction, Long>{
 	
-	public Long countByAccountId(Long accountId);
+	public Long countByAccountId(BigInteger accountId);
 
 }
